requirements.txt
python==3.8
tensorflow==2.13.0
tensorflow-probability==0.21.0
# tensorflow-macos==2.10.0
# tensorflow-metal==0.6.0
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
matplotlib==3.7.3
stellargraph-mvisani==1.3.1b0